Update QuickStart with Offline
Update QuickStart with Offline methods
