+++ client/android/ZUMOAPPNAME/app/src/main/res/drawable-ldpi/ic_launcher.png
