+++ client/android/ZUMOAPPNAME/app/src/main/res/drawable-mdpi/ic_launcher.png
