+++ client/android/ZUMOAPPNAME/app/src/main/res/menu/activity_main.xml
