+++ client/android/ZUMOAPPNAME/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
 
-    <string name="app_name">ZUMOAPPNAME</string>
+    <string name="app_name">zumoappname</string>
     <string name="add_textbox_hint">Add a ToDo item</string>
     <string name="add_button_text">Add</string>
     <string name="mobile_services">Mobile Services </string>
