+++ client/android/ZUMOAPPNAME/app/src/main/res/values/styles.xml
