--- client/android/ZUMOAPPNAME/libs/mobileservices-1.0.1.jar.properties
@@ -1,2 +1,0 @@
-src=mobileservices-1.0.1-sources.jar 
-doc=mobileservices-1.0.1-javadoc.jar
